se agregó la columna subtotal en la tabla pedidos y suma total de subtotales

+++ app/models/Producto.py
@@ -4,7 +4,7 @@ class Producto(db.Model):
     __tablename__="productos"
     id=db.Column(db.Integer, primary_key= True, autoincrement=True)
     nom_prod = db.Column(db.String(50))
-    precio = db.Column(db.Float(50))
+    precio = db.Column(db.Float())
     desc_prod = db.Column(db.String(150))
     picture_prod=db.Column(db.String(150),nullable=True)
     tipo = db.Column(db.String(50))
